+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/FlowerBoxBlock.java
@@ -6,6 +6,7 @@ import net.mehvahdjukaar.moonlight.api.misc.ForgeOverride;
 import net.mehvahdjukaar.supplementaries.common.block.ModBlockProperties;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.FlowerBoxBlockTile;
 import net.mehvahdjukaar.supplementaries.common.utils.BlockUtil;
+import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.Containers;
@@ -17,18 +18,13 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.EntityBlock;
-import net.minecraft.world.level.block.Mirror;
-import net.minecraft.world.level.block.Rotation;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.block.state.properties.BooleanProperty;
-import net.minecraft.world.level.block.state.properties.DirectionProperty;
-import net.minecraft.world.level.block.state.properties.IntegerProperty;
-import net.minecraft.world.level.material.Fluids;
+import net.minecraft.world.level.block.state.properties.*;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec3;
@@ -50,19 +46,19 @@ public class FlowerBoxBlock extends WaterBlock implements EntityBlock {
     protected static final VoxelShape SHAPE_WEST_FLOOR = Block.box(5.0D, 0.0D, 0.0D, 11.0D, 6.0D, 16.0D);
 
     public static final IntegerProperty LIGHT_LEVEL = ModBlockProperties.LIGHT_LEVEL_0_15;
-    public static final BooleanProperty FLOOR = ModBlockProperties.FLOOR;
+    public static final EnumProperty<AttachFace> ATTACHMENT = BlockStateProperties.ATTACH_FACE;
     public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
 
     public FlowerBoxBlock(Properties properties) {
         super(properties.lightLevel((s) -> s.getValue(LIGHT_LEVEL)));
         this.registerDefaultState(this.stateDefinition.any().setValue(FACING, Direction.NORTH)
-                .setValue(WATERLOGGED, false).setValue(FLOOR, false).setValue(LIGHT_LEVEL, 0));
+                .setValue(WATERLOGGED, false).setValue(ATTACHMENT, AttachFace.WALL).setValue(LIGHT_LEVEL, 0));
     }
 
     @Override
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         super.createBlockStateDefinition(builder);
-        builder.add(FACING, FLOOR, LIGHT_LEVEL);
+        builder.add(FACING, ATTACHMENT, LIGHT_LEVEL);
     }
 
     @Override
@@ -77,11 +73,32 @@ public class FlowerBoxBlock extends WaterBlock implements EntityBlock {
 
     @Override
     public BlockState getStateForPlacement(BlockPlaceContext context) {
-        boolean flag = context.getLevel().getFluidState(context.getClickedPos()).getType() == Fluids.WATER;
-        var dir = context.getClickedFace();
-        return this.defaultBlockState().setValue(WATERLOGGED, flag)
-                .setValue(FLOOR, dir == Direction.UP)
+        Direction dir = context.getClickedFace();
+        AttachFace attachface = switch (dir) {
+            case UP -> AttachFace.FLOOR;
+            case DOWN -> AttachFace.CEILING;
+            default -> AttachFace.WALL;
+        };
+        BlockState blockstate = super.getStateForPlacement(context);
+        return blockstate.setValue(ATTACHMENT, attachface)
                 .setValue(FACING, dir.getAxis().isVertical() ? context.getHorizontalDirection().getOpposite() : dir);
+    }
+
+    @Override
+    public boolean canSurvive(BlockState state, LevelReader level, BlockPos pos) {
+        Direction dir = switch (state.getValue(ATTACHMENT)) {
+            case FLOOR -> Direction.DOWN;
+            case CEILING -> Direction.UP;
+            default -> state.getValue(FACING).getOpposite();
+        };
+
+        return level.getBlockState(pos.relative(dir)).isSolid();
+    }
+
+    @Override
+    public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState, LevelAccessor worldIn, BlockPos currentPos, BlockPos facingPos) {
+        if (!this.canSurvive(stateIn, worldIn, currentPos)) return Blocks.AIR.defaultBlockState();
+        return super.updateShape(stateIn, facing, facingState, worldIn, currentPos, facingPos);
     }
 
     @Override
@@ -95,6 +112,7 @@ public class FlowerBoxBlock extends WaterBlock implements EntityBlock {
     }
 
     private static int getIndex(BlockState state, BlockPos pos, BlockHitResult hit) {
+        if (CommonConfigs.Building.FLOWER_BOX_SIMPLE_MODE.get()) return 1;
         int ind;
 
         Direction dir = state.getValue(FACING);
@@ -134,12 +152,12 @@ public class FlowerBoxBlock extends WaterBlock implements EntityBlock {
 
     @Override
     public VoxelShape getShape(BlockState state, BlockGetter world, BlockPos pos, CollisionContext context) {
-        boolean wall = !state.getValue(FLOOR);
+        boolean wall = state.getValue(ATTACHMENT) == AttachFace.WALL;
         return switch (state.getValue(FACING)) {
-            default -> wall ? SHAPE_NORTH : SHAPE_NORTH_FLOOR;
             case SOUTH -> wall ? SHAPE_SOUTH : SHAPE_NORTH_FLOOR;
             case EAST -> wall ? SHAPE_EAST : SHAPE_WEST_FLOOR;
             case WEST -> wall ? SHAPE_WEST : SHAPE_WEST_FLOOR;
+            default -> wall ? SHAPE_NORTH : SHAPE_NORTH_FLOOR;
         };
     }